Hardware and mqtt
Created the hardware folder for the sensor. Now it can detect distance and send json to MQTT every 0.5 (i think...) seconds, currently using mosquitto. Also created the mqtt docker thing, can take in the data from the sensor but still need to create database and query to tell the sensor to collect data instead of periodically.