Updated CCP and mqtt_receiver
Sensor:
change the json data to also include doc['rst'] to keep track of how many toilet paper is being used

button press will update the d_max, which is the distance when the toilet paper is at 100% capacity. Pressing this button will also send the readings to mosquitto

Instead of sending sensor reading periodically, it will now send when it is called from mqtt. The call uses callback because it is asynchornous (i think) so even it the message comes during the delay() it will still receive the call

MQTT App:
Update so that it can work with the new json format. Have equations to change raw sensor data to percentage of tissue used. Initialize the dictionary for the sensor name and floor and room that sensor name corresponds to. Also have the call function thing to have the sensor send out data every 20 minutes from 6 am to 6 pm